+++ Maquillaje/Maquillaje.WebUI/Controllers/ProveedoresController.cs
@@ -2,6 +2,7 @@
 using Maquillaje.BusinessLogic.Services;
 using Maquillaje.Entities.Entities;
 using Maquillaje.WebUI.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -25,6 +26,14 @@ namespace Maquillaje.WebUI.Controllers
         #region Detalles
         public IActionResult Details(int id)
         {
+            try
+            {
+                if (HttpContext.Session.GetString("usu_Nombre") != null)
+                {
+                    ViewBag.usu_Nombre = HttpContext.Session.GetString("usu_Nombre");
+                    ViewBag.suc_Descripcion = HttpContext.Session.GetString("suc_Descripcion");
+                    ViewBag.usu_Id = HttpContext.Session.GetString("usu_ID");
+                    ViewBag.suc_Id = HttpContext.Session.GetString("suc_Id");
 
             var cates = _generalesService.DetallesProveedores(id);
             ProveedoresViewModel model = new ProveedoresViewModel();
@@ -54,6 +63,18 @@ namespace Maquillaje.WebUI.Controllers
                 return RedirectToAction("Index"); // acá vamos a redireccionar a la pagina 404
             }
             return View(model);
+                }
+
+                return RedirectToAction("Index", "Login");
+
+
+            }
+            catch (Exception)
+            {
+                return RedirectToAction("Index", "Home");
+                throw;
+            }
+
 
 
 
@@ -64,6 +85,16 @@ namespace Maquillaje.WebUI.Controllers
         [HttpGet("/Proveedores/Listado")]
         public IActionResult Index()
         {
+            try
+            {
+                if (HttpContext.Session.GetString("usu_Nombre") != null)
+                {
+                    ViewBag.usu_Nombre = HttpContext.Session.GetString("usu_Nombre");
+                    ViewBag.suc_Descripcion = HttpContext.Session.GetString("suc_Descripcion");
+                    ViewBag.usu_Id = HttpContext.Session.GetString("usu_ID");
+                    ViewBag.suc_Id = HttpContext.Session.GetString("suc_Id");
+
+
             var listado = _generalesService.ListadoProveedores();
             var listadoMapeado = _mapper.Map<IEnumerable<ProveedoresViewModel>>(listado);
             if (TempData.ContainsKey("SuccessMessage"))
@@ -81,6 +112,20 @@ namespace Maquillaje.WebUI.Controllers
                 ViewBag.SuccessMessageScript = null;
             }
             return View(listadoMapeado);
+
+                }
+
+                return RedirectToAction("Index", "Login");
+
+
+            }
+            catch (Exception)
+            {
+                return RedirectToAction("Index", "Home");
+                throw;
+            }
+
+
         }
 
         #endregion
@@ -109,7 +154,16 @@ namespace Maquillaje.WebUI.Controllers
         [HttpPost]
         public IActionResult Create(ProveedoresViewModel item)
         {
-            if(ModelState.IsValid)
+            try
+            {
+                if (HttpContext.Session.GetString("usu_Nombre") != null)
+                {
+                    ViewBag.usu_Nombre = HttpContext.Session.GetString("usu_Nombre");
+                    ViewBag.suc_Descripcion = HttpContext.Session.GetString("suc_Descripcion");
+                    ViewBag.usu_Id = HttpContext.Session.GetString("usu_ID");
+                    ViewBag.suc_Id = HttpContext.Session.GetString("suc_Id");
+                    int usucrea = int.Parse(HttpContext.Session.GetString("usu_ID"));
+            if (ModelState.IsValid)
             {
                 if(item.prv_NombreCompañia != null && item.prv_NombreContacto != null && item.prv_DireccionEmpresa != null
                     && item.prv_DireccionContacto != null && item.prv_TelefonoContacto != null && item.prv_SexoContacto != null)
@@ -117,7 +171,7 @@ namespace Maquillaje.WebUI.Controllers
 
                     // QUITAR EL 1
                     _generalesService.CreateProveedor(item.prv_NombreCompañia,item.prv_NombreContacto, item.prv_TelefonoContacto,
-                      item.prv_DireccionEmpresa, item.prv_DireccionContacto, item.prv_SexoContacto, 1);
+                      item.prv_DireccionEmpresa, item.prv_DireccionContacto, item.prv_SexoContacto, usucrea);
                     TempData["SuccessMessage"] = "El proceso se completó correctamente";
                     MostrarToastDeExito();
                     return RedirectToAction("Index");
@@ -131,6 +185,18 @@ namespace Maquillaje.WebUI.Controllers
             {
                 return View(item);
             }
+                }
+
+                return RedirectToAction("Index", "Login");
+
+
+            }
+            catch (Exception)
+            {
+                return RedirectToAction("Index", "Home");
+                throw;
+            }
+
             
         }
 
@@ -140,6 +206,14 @@ namespace Maquillaje.WebUI.Controllers
         [HttpGet]
         public IActionResult Edit(int? id)
         {
+            try
+            {
+                if (HttpContext.Session.GetString("usu_Nombre") != null)
+                {
+                    ViewBag.usu_Nombre = HttpContext.Session.GetString("usu_Nombre");
+                    ViewBag.suc_Descripcion = HttpContext.Session.GetString("suc_Descripcion");
+                    ViewBag.usu_Id = HttpContext.Session.GetString("usu_ID");
+                    ViewBag.suc_Id = HttpContext.Session.GetString("suc_Id");
             var proveedor = _generalesService.ObtenerProveedores(id);
             if(proveedor == null)
             {
@@ -158,6 +232,19 @@ namespace Maquillaje.WebUI.Controllers
 
                 return View(item);
             }
+
+                }
+
+                return RedirectToAction("Index", "Login");
+
+
+            }
+            catch (Exception)
+            {
+                return RedirectToAction("Index", "Home");
+                throw;
+            }
+
         }
 
 
@@ -165,28 +252,54 @@ namespace Maquillaje.WebUI.Controllers
 
         public IActionResult Edit(ProveedoresViewModel item)
         {
-            if(ModelState.IsValid)
+            try
             {
+                if (HttpContext.Session.GetString("usu_Nombre") != null)
+                {
+                    ViewBag.usu_Nombre = HttpContext.Session.GetString("usu_Nombre");
+                    ViewBag.suc_Descripcion = HttpContext.Session.GetString("suc_Descripcion");
+                    ViewBag.usu_Id = HttpContext.Session.GetString("usu_ID");
+                    ViewBag.suc_Id = HttpContext.Session.GetString("suc_Id");
 
-                if(item.prv_NombreCompañia != null && item.prv_NombreContacto != null && item.prv_DireccionEmpresa != null
-                    && item.prv_DireccionContacto != null && item.prv_TelefonoContacto != null && item.prv_SexoContacto != null)
-                {
-                    // REMOVER EL 1
-                    _generalesService.UpdateProveedor(item.prv_ID, item.prv_NombreCompañia, item.prv_NombreContacto,
-                        item.prv_TelefonoContacto, item.prv_DireccionEmpresa, item.prv_DireccionContacto, item.prv_SexoContacto, 1);
-                    TempData["SuccessMessage"] = "El proceso se completó correctamente";
-                    MostrarToastDeExito();
-                    return RedirectToAction("Index");
+                    int usumodi = int.Parse(HttpContext.Session.GetString("usu_ID"));
+
+                        if (ModelState.IsValid)
+                        {
+
+                            if(item.prv_NombreCompañia != null && item.prv_NombreContacto != null && item.prv_DireccionEmpresa != null
+                                && item.prv_DireccionContacto != null && item.prv_TelefonoContacto != null && item.prv_SexoContacto != null)
+                            {
+                                // REMOVER EL 1
+                                _generalesService.UpdateProveedor(item.prv_ID, item.prv_NombreCompañia, item.prv_NombreContacto,
+                                    item.prv_TelefonoContacto, item.prv_DireccionEmpresa, item.prv_DireccionContacto, item.prv_SexoContacto, usumodi);
+                                TempData["SuccessMessage"] = "El proceso se completó correctamente";
+                                MostrarToastDeExito();
+                                return RedirectToAction("Index");
+                            }
+                            else
+                            {
+                                return View();
+                            }    
+                        }
+                        else
+                        {
+                            return View();
+                        }
+
                 }
-                else
-                {
-                    return View();
-                }    
+
+                return RedirectToAction("Index", "Login");
+
+
             }
-            else
+            catch (Exception)
             {
-                return View();
+                return RedirectToAction("Index", "Home");
+                throw;
             }
+
+
+
         }
             #endregion
 
@@ -194,16 +307,39 @@ namespace Maquillaje.WebUI.Controllers
             [HttpPost("/Proveedores/Eliminar/")]
         public IActionResult Delete(int prv_Id)
         {
+            try
+            {
+                if (HttpContext.Session.GetString("usu_Nombre") != null)
+                {
+                    ViewBag.usu_Nombre = HttpContext.Session.GetString("usu_Nombre");
+                    ViewBag.suc_Descripcion = HttpContext.Session.GetString("suc_Descripcion");
+                    ViewBag.usu_Id = HttpContext.Session.GetString("usu_ID");
+                    ViewBag.suc_Id = HttpContext.Session.GetString("suc_Id");
+                    int usumodi = int.Parse(HttpContext.Session.GetString("usu_ID"));
 
-            tbProveedores prov = new tbProveedores();
-            prov.prv_ID = prv_Id;
+                        tbProveedores prov = new tbProveedores();
+                        prov.prv_ID = prv_Id;
+                        prov.prv_UsuarioModi = usumodi;
 
 
-            var prove = _mapper.Map<tbProveedores>(prov);
-            var result = _generalesService.DeleteProv(prove);
+                        var prove = _mapper.Map<tbProveedores>(prov);
+                        var result = _generalesService.DeleteProv(prove);
 
 
-            return RedirectToAction("Index");
+                        return RedirectToAction("Index");
+
+                }
+
+                return RedirectToAction("Index", "Login");
+
+
+            }
+            catch (Exception)
+            {
+                return RedirectToAction("Index", "Home");
+                throw;
+            }
+
         }
         #endregion
